Renamed HttpClient classes to ApacheHttpClient
Removed unused StringReader class

Removed unused host parameter from HttpClients constructors

Changed access modifiers from protected to public in Module class

Took 42 minutes
