+++ Manager/src/main/java/io/raspberrywallet/manager/common/http/ApacheHttpClient.java
@@ -1,58 +1,24 @@
 package io.raspberrywallet.manager.common.http;
 
-import org.apache.commons.lang.NotImplementedException;
-import org.apache.http.Header;
 import org.apache.http.HttpResponse;
 import org.apache.http.NameValuePair;
 import org.apache.http.client.fluent.Form;
-import org.apache.http.client.methods.HttpPost;
-import org.apache.http.entity.StringEntity;
-import org.apache.http.impl.client.HttpClients;
-import org.apache.http.message.BasicHeader;
 import org.json.JSONObject;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 
-public class ApacheHttpClient extends HttpClient {
+public abstract class ApacheHttpClient {
+
+    public abstract HttpResponse sendPOSTRequest(Form body, String endpoint) throws IOException;
     
-    private String host;
-    private Header[] defaultHeaders;
+    public abstract HttpResponse sendGETRequest(Form body, String endpoint);
     
-    public ApacheHttpClient(String host, Form defaultHeaders) {
-        this.host = host;
-        this.defaultHeaders = toHeadersArray(defaultHeaders);
+    JSONObject convert(Form body) {
+        List<NameValuePair> keyValuesList = body.build();
+        JSONObject convertedJson = new JSONObject();
+        keyValuesList.forEach(pair -> convertedJson.put(pair.getName(), pair.getValue()));
+        return convertedJson;
     }
     
-    @Override
-    public HttpResponse sendPOSTRequest(Form body, String endpoint) throws IOException {
-        org.apache.http.client.HttpClient client = HttpClients.createDefault();
-        
-        JSONObject jsonObject = convert(body);
-        StringEntity stringEntity = new StringEntity(jsonObject.toString());
-        
-        HttpPost httpRequest = new HttpPost(endpoint);
-        httpRequest.setEntity(stringEntity);
-        httpRequest.setHeaders(defaultHeaders);
-        
-        return client.execute(httpRequest);
-    }
-    
-    @Override
-    public HttpResponse sendGETRequest(Form body, String endpoint) {
-        throw new NotImplementedException();
-    }
-    
-    private Header[] toHeadersArray(Form form) {
-        List<NameValuePair> list = form.build();
-        List<Header> result = new ArrayList<>(list.size());
-        list.forEach(nameValuePair -> {
-            BasicHeader basicHeader = new BasicHeader(nameValuePair.getName(), nameValuePair.getValue());
-            result.add(basicHeader);
-        });
-        
-        Header[] headers = new Header[result.size()];
-        return result.toArray(headers);
-    }
 }
